requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==4.2.27
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
whitenoise==6.11.0
